requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.12.5
authlib==1.6.9
httpx==0.28.1
itsdangerous==2.2.0
starlette==0.52.1
